file: avoid using type "long int" for reading/writing config data
Type "long int" is ambigious. It is 32 bits wide on both 32-bit and
64-bit Windows, but 64 bits wide on 64-bit Unix. Since we read and
write 4 bytes in File::loadInt() and File::storeInt(), respectively,
we want a type which is 32 bits wide on all platforms. Thus, use plain
"int" instead of "long int". I don't know of any platform on which int
has more or less than 32 bits, especially none still supported by SDL.

+++ src/io/file.cpp
@@ -273,29 +273,29 @@ void File::storeShort (unsigned short int val) {
 
 
 /**
- * Load a signed long int from the file.
+ * Load a signed int from the file.
  *
  * @return The value read
  */
-signed long int File::loadInt () {
+signed int File::loadInt () {
 
-	unsigned long int val;
+	unsigned int val;
 
 	val = fgetc(file);
 	val += fgetc(file) << 8;
 	val += fgetc(file) << 16;
 	val += fgetc(file) << 24;
 
-	return *((signed long int *)&val);
+	return *((signed int *)&val);
 
 }
 
 
-void File::storeInt (signed long int val) {
+void File::storeInt (signed int val) {
 
-	unsigned long int uval;
+	unsigned int uval;
 
-	uval = *((unsigned long int *)&val);
+	uval = *((unsigned int *)&val);
 
 	fputc(uval & 255, file);
 	fputc((uval >> 8) & 255, file);
